Split switch estate into own controller
Also some other minor refactorings.
