+++ config/routes.rb
@@ -53,9 +53,9 @@ Rails.application.routes.draw do
       get '/processed', action: :processed, as: 'processed_visits'
       get '/print_visits', action: :print_visits, as: 'print_visits'
       get '/search', action: :search, as: :search
-      post '/switch_estates', action: :switch_estates, as: 'switch_estates'
     end
 
+    resource :switch_estates, only: %i[ create ]
     resources :feedbacks, only: %i[ new create ]
   end
